demo-app/service1/requirements.txt
Flask==1.1.2
python-json-logger==2.0.1
uwsgi
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp
grpcio
opencensus-proto
opentelemetry-api
opentelemetry-sdk

